Move the functions for cleaning to minicriu-client.c
